+++ predict.h
@@ -5,7 +5,7 @@
 
 int prediction_module_init(int file_counter);
 void predict_init_wait(void);
-long long int get_predict_init_time(void);
+long int get_predict_init_time(void);
 void stop_prediction_thr(void);
 void set_predict_request_aggregation_flag(int value);
 void lock_prediction_thr_refresh_mutex(void);
@@ -19,11 +19,11 @@ void inc_current_predicted_reqfilenb(void);
 int get_current_predicted_reqfilenb(void);
 
 void prediction_newreq(struct request_t *req);
-long long int agios_predict_should_we_wait(struct request_t *req);
+long int agios_predict_should_we_wait(struct request_t *req);
 int predict_select_best_algorithm(void);
 
 void calculate_average_stripe_access_time_difference(struct related_list_t *related_list); 
-void update_average_distance(struct related_list_t *related_list, long long offset, long len);
+void update_average_distance(struct related_list_t *related_list, long offset, long len);
